Updated IRQ request to avoid IRQ loosing
On some condition, when driver initialize, IRQ is forgotten/deleted by system. Power switch is no longer useable (no powerdown when clicking. Kernel detect clicking, but can send an IRQ as it does not longer exist). Main cause is IRQ request is not link with a cookie, allowing any to delete it. When setting a cookie, everything works fine.